fix(responses): move model file id helpers to core
