+++ litellm/proxy/openai_files_endpoints/common_utils.py
@@ -5,6 +5,11 @@ from dataclasses import dataclass, field
 from types import MappingProxyType
 from typing import TYPE_CHECKING, List, Literal, Optional, Union
 
+from litellm.litellm_core_utils.file_id_utils import (
+    decode_model_from_file_id,
+    get_original_file_id,
+    is_model_embedded_id,  # noqa: F401  # preserve the proxy-module import path
+)
 from litellm.repositories.table_repositories import (
     ManagedFileRepository,
     ManagedObjectRepository,
@@ -147,72 +152,6 @@ def encode_batch_response_ids(response, model: str) -> None:
                 attr,
                 encode_file_id_with_model(file_id=getattr(response, attr), model=model),
             )
-
-
-def decode_model_from_file_id(encoded_id: str) -> Optional[str]:
-    """
-    Extract model name from an encoded file/batch ID.
-    Handles IDs that start with "file-" or "batch_" prefix.
-    """
-    try:
-        if not isinstance(encoded_id, str):
-            return None
-
-        # Remove prefix if present (file-, batch_, etc.)
-        if encoded_id.startswith("file-"):
-            b64_part = encoded_id[5:]  # Remove "file-"
-        elif encoded_id.startswith("batch_"):
-            b64_part = encoded_id[6:]  # Remove "batch_"
-        else:
-            b64_part = encoded_id
-
-        padded = b64_part + "=" * (-len(b64_part) % 4)
-        decoded = base64.urlsafe_b64decode(padded).decode()
-        if decoded.startswith("litellm:") and ";model," in decoded:
-            match = re.search(r";model,([^;]+)", decoded)
-            if match:
-                return match.group(1).strip()
-
-        return None
-    except Exception:
-        return None
-
-
-def get_original_file_id(encoded_id: str) -> str:
-    """
-    Extract the original provider file/batch ID from an encoded ID.
-    Handles IDs that start with "file-" or "batch_" prefix.
-    """
-    try:
-        if not isinstance(encoded_id, str):
-            return encoded_id
-
-        # Remove prefix if present (file-, batch_, etc.)
-        if encoded_id.startswith("file-"):
-            b64_part = encoded_id[5:]  # Remove "file-"
-        elif encoded_id.startswith("batch_"):
-            b64_part = encoded_id[6:]  # Remove "batch_"
-        else:
-            b64_part = encoded_id
-
-        padded = b64_part + "=" * (-len(b64_part) % 4)
-        decoded = base64.urlsafe_b64decode(padded).decode()
-
-        if decoded.startswith("litellm:") and ";model," in decoded:
-            match = re.search(r"litellm:([^;]+);model,", decoded)
-            if match:
-                return match.group(1)
-
-        return encoded_id
-    except Exception:
-        return encoded_id
-
-
-def is_model_embedded_id(file_id: str) -> bool:
-    """
-    Check if a file/batch ID has model routing information embedded.
-    """
-    return decode_model_from_file_id(file_id) is not None
 
 
 # ============================================================================
